HMAN-33 tidy up tests
